fixed donothing and tweaked autos

+++ src/Commands/Other/DoNothing.cpp
@@ -1,6 +1,6 @@
 #include <Commands/Other/DoNothing.h>
 
-DoNothing::DoNothing(int input)
+DoNothing::DoNothing(double input)
 {
 	timer = input;
 }
